requirements.txt
setuptools==62.3.2
django>=2.2.1,<=3.2.13
graphene-django==2.15.0
graphene-file-upload==1.2.2
pyjwt==2.3.0
urllib3==1.26.9
pillow==8.4.0
facebook-sdk==3.1.0
google-auth==1.6.3
moodlepy==0.22.2
django-auditor-logs==1.0.1
django-extensions==3.1.3